requirements.txt
streamlit
fastapi
pandas
openpyxl
plotly
transformers
uvicorn
requests
openai
python-dotenv
torch
numpy
matplotlib
seaborn
pytest
torch